+++ frontend/src/models/levelsets/buffers.tsx
@@ -30,6 +30,7 @@ import {
   IBaseLevel,
   IBaseLevelset,
   levelSupportsDemo,
+  LocalOptions,
   summarySupportReport,
   SupportReportType,
 } from "drivers";
@@ -111,7 +112,11 @@ fileDidOpen.watch((key) => {
   setCurrentLevel(0);
 });
 export const exportCurrentLevel = createEvent<any>();
-export const importCurrentLevel = createEvent<File>();
+interface ImportLevelParams {
+  file: File;
+  levelset?: IBaseLevelset<IBaseLevel>;
+}
+export const importCurrentLevel = createEvent<ImportLevelParams>();
 /**
  * Insert new level in current level index in current levelset
  */
@@ -977,7 +982,12 @@ sample({
   }
 
   const importCurrentLevelFx = createEffect(
-    async ({ file, driverName, driverFormat }: { file: File } & DriverOpt) => {
+    async ({
+      file,
+      levelset,
+      driverName,
+      driverFormat,
+    }: ImportLevelParams & DriverOpt) => {
       const { formats } = getDriver(driverName)!;
       const { createLevelset, readLevelset, supportReport, writeLevelset } =
         formats[driverFormat];
@@ -992,7 +1002,12 @@ sample({
       }
       const from = formats[fromFormat];
 
-      let levelset = from.readLevelset(await file.arrayBuffer());
+      let localOptions: LocalOptions | undefined = undefined;
+      if (levelset) {
+        localOptions = levelset.getLevel(0).localOptions;
+      } else {
+        levelset = from.readLevelset(await file.arrayBuffer());
+      }
       const report = summarySupportReport(supportReport(levelset));
       if (report?.type === SupportReportType.ERR) {
         return { report };
@@ -1001,7 +1016,10 @@ sample({
       levelset = readLevelset(
         writeLevelset(createLevelset([levelset.getLevel(0)])),
       );
-      return { level: levelset.getLevel(0), report };
+      return {
+        level: levelset.getLevel(0).setLocalOptions(localOptions),
+        report,
+      };
     },
   );
   sample({
@@ -1011,7 +1029,7 @@ sample({
       driverFormat: $currentDriverFormat,
     },
     filter: (o): o is DriverOpt => Boolean(o.driverName && o.driverFormat),
-    fn: (s: DriverOpt, file: File) => ({ ...s, file }),
+    fn: (s: DriverOpt, f: ImportLevelParams) => ({ ...s, ...f }),
     target: importCurrentLevelFx,
   });